move formatTimeAgo
